避免重复进入bool mode，将文字hash映射到int128上，避免重复加入子句

+++ src/smt/nia_ls/nia_build_instance.cpp
@@ -349,20 +349,18 @@ void ls_solver::unit_prop(){
         }
     }
 }
-int ls_solver::hash_lits_to_num(std::vector<int> &lits){
+__int128_t ls_solver::hash_lits_to_num(std::vector<int> &lits){
     std::sort(lits.begin(), lits.end());
-    int hash_num=0;
-    for(int lit_idx:lits){
-        hash_num=hash_num*(int)(_num_lits)+lit_idx+(int)_num_lits;
-    }
+    __int128_t hash_num=0;
+    for(int lit_idx:lits){hash_num=(__int128_t)hash_num*(__int128_t)(_num_lits)+(__int128_t)lit_idx+(__int128_t)_num_lits;}
     return hash_num;
 }
 
 void ls_solver::resolution(){
     std::vector<uint64_t> pos_clauses(10*_num_clauses);
     std::vector<uint64_t> neg_clauses(10*_num_clauses);
-    std::map<int,int>  clauselit_map;//for the clause with literal {a,b,c}, sort the lit by its order, and hash the literals to a number, then map it to the clause_idx, if deleted, set it to -1
-    std::vector<int>    clauselit(_clauses.size());//hash the lits of clause to a number
+    std::map<__int128_t,int>  clauselit_map;//for the clause with literal {a,b,c}, sort the lit by its order, and hash the literals to a number, then map it to the clause_idx, if deleted, set it to -1
+    std::vector<__int128_t>    clauselit(_clauses.size());//hash the lits of clause to a number
     for(int cls_idx=0;cls_idx<_clauses.size();cls_idx++){
         clauselit[cls_idx]=hash_lits_to_num(_clauses[cls_idx].literals);
         clauselit_map[clauselit[cls_idx]]=cls_idx;
@@ -449,7 +447,7 @@ void ls_solver::resolution(){
                         if(!is_existed_lit){new_clause.literals.push_back(l_sign_idx);}
                     }
                 }
-                int clause_lit_hash=hash_lits_to_num(new_clause.literals);
+                __int128_t clause_lit_hash=hash_lits_to_num(new_clause.literals);
                 if(!is_tautology&&(clauselit_map.find(clause_lit_hash)==clauselit_map.end()||clauselit_map[clause_lit_hash]==-1)){//add new clause, and modify the clause of corresponding bool var
                     for(int l_sign_idx:new_clause.literals){
                         lit *l_inner=&(_lits[std::abs(l_sign_idx)]);
